Add footer for -a with avg and stddev of p-values
as deviation from the ideal p-value per test: abs(p-value - 0.5).
0.0 is best, 0.5 worst.
This way have an overall quality number for an rng.

+++ dieharder/dieharder.h
@@ -64,7 +64,7 @@ EXTERN double strategy;
  void time_rng();
  void set_globals();
  void choose_rng();
- int execute_test(int);
+ double execute_test(int);
  void run_all_tests();
  void run_test();
  void add_ui_rngs();
